requirements.txt
Flask==1.1.2
webapp3-flask==0.1.9
google-cloud-storage
google-auth-oauthlib==0.4.4
google-api-python-client==2.6.0
google-auth==1.30.1
google-cloud-ndb==1.7.1
MonthDelta==0.9.1
grpcio==1.29.0
xlrd==1.2.0
plivo==4.14.0
google-cloud==0.34.0
requests-oauthlib==1.3.0
requests==2.26.0
pytz==2021.1